Command should also be an Action
